requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pdf2image==1.16.3
Pillow==10.0.1
PyMuPDF==1.23.4
PyMuPDFb==1.23.3
pytesseract==0.3.10
Werkzeug==2.3.7
